+++ src/libcsg/modules/io/dlpolytopologyreader.cc
@@ -61,7 +61,7 @@ string DLPOLYTopologyReader::_NextKeyline(ifstream &fs, const char *wspace)
 }
 
 string DLPOLYTopologyReader::_NextKeyInt(ifstream &fs, const char *wspace,
-                                         const string &word, long int &ival)
+                                         const string &word, long &ival)
 // function to read the next line containing only a given keyword and an integer
 // value after it (only skipping comments!) NOTE: this function must only be
 // called when the next directive line has to contain the given keyword and an
@@ -140,8 +140,8 @@ bool DLPOLYTopologyReader::_isKeyInt(const string &line, const char *wspace,
 bool DLPOLYTopologyReader::ReadTopology(string file, Topology &top) {
   const char *WhiteSpace = " \t";
 
-  long int matoms = 0;
-  long int natoms = 0;
+  long matoms = 0;
+  long natoms = 0;
 
   std::ifstream fl;
   boost::filesystem::path filepath(file.c_str());
@@ -203,7 +203,7 @@ bool DLPOLYTopologyReader::ReadTopology(string file, Topology &top) {
       mol_name = _NextKeyline(fl, WhiteSpace);
       Molecule *mi = top.CreateMolecule(mol_name);
 
-      long int nreplica = 1;
+      long nreplica = 1;
       line = _NextKeyInt(fl, WhiteSpace, "NUMMOL", nreplica);
 
 #ifdef DEBUG
@@ -219,7 +219,7 @@ bool DLPOLYTopologyReader::ReadTopology(string file, Topology &top) {
 #endif
 
       // read molecule
-      std::vector<int> id_map(natoms);
+      std::vector<long> id_map(natoms);
       for (int i = 0; i < natoms;) {  // i is altered in repeater loop
         stringstream sl(_NextKeyline(fl, WhiteSpace));
 
@@ -353,7 +353,7 @@ bool DLPOLYTopologyReader::ReadTopology(string file, Topology &top) {
         InteractionContainer ics = mi->Interactions();
         for (auto &ic : ics) {
           Interaction *ic_replica = nullptr;
-          long int offset =
+          long offset =
               mi_replica->getBead(0)->getId() - mi->getBead(0)->getId();
           if (ic->BeadCount() == 2) {
             ic_replica =
